Move platform window management to window.rs files
